refactor(init): move friendly_path to the module that owns paths
Three modules call it — the receipt, the config-rewrite prompt and the reload
prompt — so it belongs beside the other path helpers rather than inside the
renderer that happened to be written first. Its test moves with it and is
named for the behavior it checks.
